refactor(#1142): address code and architecture review findings
- Extract shared ffmpeg WAV conversion to FfmpegAudioConverter; both
  AzureSpeechTranscriptionService and WhisperTranscriptionService now
  delegate to it instead of duplicating the pipeline
- Hoist Transcription:Provider read above both validation and DI blocks
  so a single variable drives both without risk of default drift
- Add Win32Exception catch around Process.Start and surface ffmpeg
  stderr in the thrown InvalidOperationException
- Log warning when Whisper response lacks a 'text' property
- openai.bicep: add appPrincipalId param and vault-level role assignment
  (idempotent with keyvault.bicep grant) for module self-containment
- required-secrets.json: clarify mutual-exclusivity of provider secrets
- Tighten test assertion to InvalidOperationException
