+++ infra/required-secrets.json
@@ -1,5 +1,5 @@
 {
-  "_readme": "Secret names use double-dash (--) as the hierarchy separator, mirroring ASP.NET Core Key Vault config mapping (colon -> double-dash). These must match the keys in StartupConfigValidator (backend/LangTeach.Api/Infrastructure/StartupConfigValidator.cs) and the secrets defined in infra/modules/keyvault.bicep.",
+  "_readme": "Secret names use double-dash (--) as the hierarchy separator, mirroring ASP.NET Core Key Vault config mapping (colon -> double-dash). These must match the keys in StartupConfigValidator (backend/LangTeach.Api/Infrastructure/StartupConfigValidator.cs) and the secrets defined in infra/modules/keyvault.bicep. NOTE: AzureSpeech--* and AzureOpenAIWhisper--* are mutually exclusive at runtime — only the set matching Transcription:Provider is validated at startup. Both sets appear here so operators know which secrets to provision for each provider.",
   "requiredSecrets": [
     "ConnectionStrings--Default",
     "Auth0--Domain",
